+++ src/tests/outputs/postgres.output.ts
@@ -7,10 +7,10 @@ export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
 export type Status = "CONFIRMED" | "UNCONFIRMED";
 
 export interface UserTest {
-  array: string[] | null;
   id: Generated<number>;
   userStatus: Status | null;
   userStatus2: Status | null;
+  array: string[] | null;
 }
 
 export interface DB {