+++ src/nodes/literal-node.ts
@@ -1,10 +1,12 @@
 import { NodeType } from '../enums';
 
-export class LiteralNode {
-  readonly type = NodeType.LITERAL;
-  readonly value: string;
+export type Literal = number | string;
 
-  constructor(value: string) {
+export class LiteralNode<T extends Literal = Literal> {
+  readonly type = NodeType.LITERAL;
+  readonly value: T;
+
+  constructor(value: T) {
     this.value = value;
   }
 }