+++ src/tests/outputs/sqlite.output.ts
@@ -1,8 +1,16 @@
-export interface UserTest {
-  id: string | null;
+import { ColumnType } from "kysely";
+
+export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
+  ? ColumnType<S, I | undefined, U>
+  : ColumnType<T, T | undefined, T>;
+
+export interface FooBar {
+  false: number;
+  true: number;
+  id: Generated<number>;
   userStatus: string | null;
 }
 
 export interface DB {
-  userTest: UserTest;
+  fooBar: FooBar;
 }