+++ src/constants/imports.ts
@@ -2,5 +2,4 @@ import { ModuleReferenceNode } from '../nodes';
 
 export const GLOBAL_IMPORTS = {
   ColumnType: new ModuleReferenceNode('kysely'),
-  RawBuilder: new ModuleReferenceNode('kysely'),
 };